move CommandExecutor to infrastructure package
