+++ src/main/scala/swiss/dasch/domain/MovingImageService.scala
@@ -1,12 +1,10 @@
 package swiss.dasch.domain
 
 import org.apache.commons.io.FilenameUtils
-import swiss.dasch.config.Configuration.SipiConfig
 import swiss.dasch.domain.DerivativeFile.MovingImageDerivativeFile
+import swiss.dasch.infrastructure.CommandExecutor
 import zio.json.{DecoderOps, DeriveJsonDecoder, JsonDecoder}
 import zio.{Task, ZIO, ZLayer}
-
-import scala.sys.process.{ProcessLogger, stringToProcess}
 
 final case class MovingImageMetadata(width: Int, height: Int, duration: Double, fps: Int)
 
@@ -64,48 +62,4 @@ case class MovingImageService(storage: StorageService, executor: CommandExecutor
 
 object MovingImageService {
   val layer = CommandExecutor.layer >>> ZLayer.derive[MovingImageService]
-}
-
-final case class Cmd(cmd: String)
-final case class CommandExecutor(sipiConfig: SipiConfig, storageService: StorageService) {
-
-  def buildCommand(command: String, params: String): Task[Cmd] =
-    for {
-      assetDir <- storageService.getAssetDirectory().flatMap(_.toAbsolutePath)
-    } yield
-      if (sipiConfig.useLocalDev) {
-        Cmd(s"docker run --entrypoint $command -v $assetDir:$assetDir daschswiss/knora-sipi:latest $params")
-      } else {
-        Cmd(s"$command $params")
-      }
-
-  private case class ProcessOutput(out: String, err: String)
-
-  private class InMemoryProcessLogger extends ProcessLogger {
-    private val sbOut = new StringBuilder
-    private val sbErr = new StringBuilder
-
-    override def out(s: => String): Unit = sbOut.append(s + "\n")
-
-    override def err(s: => String): Unit = sbErr.append(s + "\n")
-
-    override def buffer[T](f: => T): T = f
-
-    def getOutput: ProcessOutput = ProcessOutput(sbOut.toString(), sbErr.toString())
-  }
-
-  def execute(command: Cmd): Task[String] = {
-    val logger = new InMemoryProcessLogger()
-    for {
-      _   <- ZIO.logInfo(s"Executing command: ${command.cmd}")
-      out <- ZIO.attemptBlockingIO(command.cmd ! logger).as(logger.getOutput)
-      _ <- ZIO.when(out.err.nonEmpty)(
-             ZIO.fail(new RuntimeException(s"Failed executing '${command.cmd}' with error '${out.err}''"))
-           )
-    } yield out.out
-  }
-}
-
-object CommandExecutor {
-  val layer = ZLayer.derive[CommandExecutor]
 }
